pyproject.toml
[project]
name = "arzw-stellar-odyssey"
version = "0.1.0"
description = ""
authors = [
    {name = "arzw",email = "dunkmemee@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<4"
dependencies = [
    "django (>=5.2,<6.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "django-ninja (>=1.4.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
isort = "^6.0.1"


[tool.isort]
multi_line_output=3
include_trailing_comma=true
line_length=120
lines_after_imports=2
lines_between_sections=1
lines_between_types=0
force_alphabetical_sort_within_sections=true
ensure_newline_before_comments=true
combine_as_imports=true
combine_star=true
skip_gitignore=true
balanced_wrapping=true
force_grid_wrap=2
src_paths=["src", ]
skip_glob=[
    ".git",
    ".idea",
    ".vscode",
    "venv",
    ".venv",
]
known_django=["django","ninja"]
known_first_party=["src"]
sections=[
    "FUTURE",
    "STDLIB",
    "DJANGO",
    "THIRDPARTY",
    "FIRSTPARTY",
    "LOCALFOLDER",
]